real*8 changed to real(r8) in obs_def_CO_Nadir_mod.
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@9027 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 7b9424abff2f46e9acfdcc50ee6a38b15ae8e3aa

+++ obs_def/obs_def_CO_Nadir_mod.f90
@@ -549,9 +549,9 @@ end subroutine get_expected_mopitt_co
 ! Allows passing of obs_def special information 
 
 integer,	 	intent(in)	:: key, co_nlevels
-real*8,dimension(10),	intent(in)	:: co_avgker	
-real*8,			intent(in)	:: co_prior
-real*8,			intent(in)	:: co_psurf
+real(r8),dimension(10),	intent(in)	:: co_avgker
+real(r8),			intent(in)	:: co_prior
+real(r8),			intent(in)	:: co_psurf
 character(len=129) 			:: msgstring
 
 if ( .not. module_initialized ) call initialize_module
